+++ src/network/networkd-dhcp6.h
@@ -2,7 +2,6 @@
 #pragma once
 
 #include "conf-parser.h"
-#include "in-addr-util.h"
 #include "macro.h"
 
 typedef enum DHCP6ClientStartMode {
@@ -17,13 +16,10 @@ typedef struct Link Link;
 typedef struct Request Request;
 
 bool link_dhcp6_with_address_enabled(Link *link);
-bool link_dhcp6_pd_is_enabled(Link *link);
-int dhcp6_pd_find_uplink(Link *link, Link **ret);
-int dhcp6_pd_remove(Link *link, bool only_marked);
+int dhcp6_check_ready(Link *link);
 int dhcp6_update_mac(Link *link);
 int dhcp6_start(Link *link);
 int dhcp6_start_on_ra(Link *link, bool information_request);
-int dhcp6_request_prefix_delegation(Link *link);
 
 int request_process_dhcp6_client(Request *req);
 int link_request_dhcp6_client(Link *link);
@@ -33,7 +29,6 @@ int link_serialize_dhcp6_client(Link *link, FILE *f);
 CONFIG_PARSER_PROTOTYPE(config_parse_dhcp6_pd_prefix_hint);
 CONFIG_PARSER_PROTOTYPE(config_parse_dhcp6_mud_url);
 CONFIG_PARSER_PROTOTYPE(config_parse_dhcp6_client_start_mode);
-CONFIG_PARSER_PROTOTYPE(config_parse_dhcp6_pd_subnet_id);
 
 const char* dhcp6_client_start_mode_to_string(DHCP6ClientStartMode i) _const_;
 DHCP6ClientStartMode dhcp6_client_start_mode_from_string(const char *s) _pure_;
